6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 /**
  * main - print out sizes of data types in C
- * code by uhokay21
+ *
  * Return: 0
  */
 int main(void)
 {
-	printf("Size of a char: %lu byte(S)\n", (unsigned long)sizeof(char));
-	printf("Size of a int: %lu byte(S)\n", (unsigned long)sizeof(int));
-	printf("Size of a long int: %lu byte(S)\n", (unsigned long)sizeof(long int));
-	printf("Size of a long long int: %lu byte(S)\n", (unsigned long)sizeof(long long int));
-	printf("Size of a float: %lu byte(S)\n", (unsigned long)sizeof(float));
+	printf("Size of a char: %lu byte(S)\n", sizeof(char));
+	printf("Size of a int: %lu byte(S)\n", sizeof(int));
+	printf("Size of a long int: %lu byte(S)\n", sizeof(long int));
+	printf("Size of a long long int: %lu byte(S)\n", sizeof(long long int));
+	printf("Size of a float: %lu byte(S)\n", sizeof(float));
 	return (0);
 }
